refactor(simulation): extract branching logic into branching_service
Issue #44 (EPIC-06-ST-04). Branch-Logik aus SimulationManager (789 LOC) in
neues Funktionsmodul backend/app/services/branching_service.py (230 LOC)
gezogen: list_branches, create_branch, _apply_persona_overrides.

- Funktionsmodul statt Service-Klasse; Manager wird als erster Parameter
  übergeben (vermeidet zirkulären Import).
- SimulationStatus per lazy import in create_branch geladen.
- Manager-Methoden sind 1-Zeilen-Delegationen — Caller-API stabil
  (api/simulation_profiles.py und Tests unverändert).
- Manager-Imports aufgeräumt: shutil, Config, ArtifactLocator, RunRegistry
  nicht mehr nötig. Manager schrumpft auf 622 LOC (-21 %).

Verifikation: npm run check 5/5 grün — Backend 517 passed (2 skipped Redis),
Frontend 40 passed, Lint clean, Build ok.

Slice A2 von v0.9.0-Pfad A.

Closes #44
